refactor: renamed find_shapes correctly
